Refactor geometry package for improved clarity and consistency
- Updated shear factors in shear.go to use ToScalar for better type handling.
- Refined Size interface in size.go for clearer method signatures and improved documentation.
- Enhanced superellipse calculations in superellipse.go, optimizing variable usage and ensuring consistent type handling.
- Streamlined vector operations in vector.go, implementing NewVector3 and NewVector4 for better readability and maintainability.
- Removed redundant cross product test for 4D vectors in vector_test.go, as it is not applicable.
- Improved cubic and quadratic subdivision calculations in wangs.go for better precision.
- Added comprehensive tests for Int26_6 point operations in geom_test.go, covering basic arithmetic, length, normalization, and string representation.

+++ geom/example/color_blend/main.go
@@ -39,17 +39,17 @@ type BlendModeDemo struct {
 
 // At generates optimized test patterns for accurate blend mode verification
 func (b *BlendModeDemo) At(x int, y int) color.Color {
-	width := float64(b.rect.Width())
-	height := float64(b.rect.Height())
+	width := int(b.rect.Width())
+	height := int(b.rect.Height())
 
 	// Grid layout: 7 columns × 5 rows for 29 blend modes (0-28)
 	cols := 7
 	rows := 5
-	cellW := width / float64(cols)
-	cellH := height / float64(rows)
+	cellW := width / cols
+	cellH := height / rows
 
-	col := int(float64(x) / cellW)
-	row := int(float64(y) / cellH)
+	col := x / cellW
+	row := y / cellH
 
 	if col >= cols {
 		col = cols - 1
@@ -67,8 +67,8 @@ func (b *BlendModeDemo) At(x int, y int) color.Color {
 	}
 
 	// Calculate normalized coordinates within cell [0,1]
-	fx := (float64(x) - float64(col)*cellW) / cellW
-	fy := (float64(y) - float64(row)*cellH) / cellH
+	fx := float64((x - col*cellW)) / float64(cellW)
+	fy := float64((y - row*cellH)) / float64(cellH)
 
 	// Reserve space for text labels at top
 	if fy < 0.25 {
